ai/requirements_captcha_learning.txt
# CAPTCHA Learning System Requirements
# For NVIDIA GPU support with CUDA

torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.0.0

# GUI and Input Handling
PyAutoGUI>=0.9.54
keyboard>=0.13.5
pynput>=1.7.6

# Data Science
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0

# Utilities
tqdm>=4.65.0
requests>=2.31.0
